requirement.txt
requests==2.31.0
python-dotenv==1.0.1
telethon==1.29.2
websocket-client